#17 - Refactor configuration classes and add YAML support for application settings

+++ app/src/main/java/com/cuseto/pong/config/BallConfig.java
@@ -2,6 +2,6 @@ package com.cuseto.pong.config;
 
 public record BallConfig(
     int radius,
-    float initialVelocityX,
-    float initialVelocityY
+    double initialVelocityX,
+    double initialVelocityY
 ) {}
